+++ Assets/OxGFrame/AssetLoader/Scripts/Editor/Bundle/CryptogramSetting.cs
@@ -2,7 +2,7 @@
 
 namespace OxGFrame.AssetLoader.Editor
 {
-    [CreateAssetMenu(fileName = nameof(CryptogramSetting), menuName = "OxGFrame/Create Settings/Cryptograme Setting")]
+    [CreateAssetMenu(fileName = nameof(CryptogramSetting), menuName = "OxGFrame/Create Settings/Cryptogram Setting")]
     public class CryptogramSetting : ScriptableObject
     {
         [Header("OFFSET")]
@@ -12,9 +12,10 @@ namespace OxGFrame.AssetLoader.Editor
         [Header("XOR")]
         public byte xorKey = 1;
 
-        [Header("HTXOR")]
+        [Header("HT2XOR")]
         public byte hXorKey = 1;
         public byte tXorKey = 1;
+        public byte jXorKey = 1;
 
         [Header("AES")]
         public string aesKey = "aes_key";